update teacher controller and reflect the required change

+++ src/main/java/com/example/demo/controller/TeacherController.java
@@ -9,18 +9,18 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import com.example.demo.entity.Teacher;
-import com.example.demo.serviceimp.TeacherImpl;
 import java.util.List;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import com.example.demo.serviceinterface.TeacherService;
 
 @RestController
 @RequestMapping("/teachers")
 public class TeacherController {
 
-    private final TeacherImpl teacherService;
+    private final TeacherService teacherService;
 
-    public TeacherController(TeacherImpl teacherService) {
+    public TeacherController(TeacherService teacherService) {
         this.teacherService = teacherService;
     }
 
@@ -37,20 +37,34 @@ public class TeacherController {
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<Teacher> update(@RequestBody Teacher teacherDetails) {
-        Teacher updatedTeacher = teacherService.update(teacherDetails);
-        return new ResponseEntity<>(updatedTeacher, HttpStatus.OK);
+    public ResponseEntity<Teacher> update(@PathVariable Long id, @RequestBody Teacher teacherDetails) {
+        // return new ResponseEntity<>(updatedTeacher, HttpStatus.OK);
+        if (id.equals(teacherDetails.getId()) && teacherService.existsById(id)) {
+            Teacher updatedTeacher = teacherService.update(teacherDetails);
+            return ResponseEntity.status(HttpStatus.OK).body(updatedTeacher);
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(null);
+        }
     }
 
     @GetMapping("/{id}")
     public ResponseEntity<Teacher> get(@PathVariable Long id) {
-        Teacher teacher = teacherService.get(id);
-        return new ResponseEntity<>(teacher, HttpStatus.OK);
+        // return new ResponseEntity<>(teacher, HttpStatus.OK);
+        if (teacherService.existsById(id)) {
+            Teacher teacher = teacherService.get(id);
+            return ResponseEntity.status(HttpStatus.OK).body(teacher);
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(null);
+        }
     }
 
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> delete(@PathVariable Long id) {
-        teacherService.delete(id);
-        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        if (teacherService.existsById(id)) {
+            teacherService.delete(id);
+            return ResponseEntity.status(HttpStatus.OK).body(null);
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(null);
+        }
     }
 }
